pyproject.toml
[tool.poetry]
name = "semi-auto-novelai-to-pixiv"
version = "2.0.0"
description = "通过 Nai3 生成图片然后经过人工筛选上传 Pixiv"
authors = ["zhulinyv <zhulinyv2005@outlook.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "semi_auto_novelai_to_pixiv"}]

[tool.poetry.dependencies]
python = "^3.10"
setuptools = "^70.0.0"
pydantic = "^2.7.3"
gradio = "^4.36.1"
pydantic-settings = "^2.3.1"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
requests = "^2.32.3"
loguru = "^0.7.2"
ujson = "^5.10.0"
nudenet = "^3.0.8"
opencv-python = "^4.9.0.80"
pynvml = "^11.5.0"
g4f = "^0.3.1.9"
pypiwin32 = "^223"
flask = "^3.0.3"
numpy = "^1.26.4"
gitpython = "^3.1.43"
moviepy = "^1.0.3"
send2trash = "^1.8.3"
playsound = "1.2.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

# https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[tool.black]
line-length = 120

# https://beta.ruff.rs/docs/settings/
[tool.ruff]
line-length = 120
# https://beta.ruff.rs/docs/rules/
lint.select = ["E", "W", "F"]
lint.ignore = ["E501"]
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
lint.ignore-init-module-imports = true

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"